+++ branches/mod/shelter/main.js
@@ -87,7 +87,7 @@ electron.ipcMain.handle("SHELTER_BRANCHCHANGE_SECURITY_DIALOG", (_, message) =>
 		buttons: ["Cancel", "Confirm"],
 		title: "Sheltupdate mods change",
 		detail:
-			'We confirm for security reasons that this action is intended by the user. Only continue if you got here from the shelter "Client Mods" UI.',
+			'For security reasons please confirm that this action is intended by the user. Only continue if you got here from the shelter "Client Mods" UI.',
 	}),
 );
 // #endregion